refactor: stage model, drop ExecuteOptions global, split ast/execute
Finish the #1384 library-first engine boundary: ExecuteOptions holds only
EngineContext + PathGuard (from_global is a CLI boundary adapter).

Unify write staging on WriteRequest / WriteSource / WriteReport via stage(),
with execute_single/operations/precomputed as thin wrappers.

Split oversized monofiles without size waivers:
- src/tx/execute/ (mod, file_ops, policy, tests)
- src/cmd/ast/ (mod, common, query, mutate)

Closes spirit of follow-ups after #1387 structural rewrite.
